Moved restivus related main APIs to /server directory
Files management.js and catalog.js moved from /apinf_packages/rest_apis
to  /apinf_packages/rest_apis/server
