+++ apinf_packages/apis/server/api.js
@@ -11,7 +11,7 @@ import { Roles } from 'meteor/alanning:roles';
 
 // Collection imports
 import Apis from '/apinf_packages/apis/collection';
-import CatalogV1 from '/apinf_packages/rest_apis/catalog';
+import CatalogV1 from '/apinf_packages/rest_apis/server/catalog';
 import Organizations from '/apinf_packages/organizations/collection';
 
 CatalogV1.swagger.meta.paths = {
