+++ apinf_packages/users/collection/server/api.js
@@ -11,7 +11,7 @@ import { Accounts } from 'meteor/accounts-base';
 import { Roles } from 'meteor/alanning:roles';
 
 // Collection imports
-import ManagementV1 from '/apinf_packages/rest_apis/management';
+import ManagementV1 from '/apinf_packages/rest_apis/server/management';
 import Organizations from '/apinf_packages/organizations/collection';
 
 // Npm packages imports
